+++ packages/pluggableWidgets/tree-node-web/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+-   We fixed an issue with nested Tree Nodes, where the nested empty Tree Node would break its parent behavior.
+
 ## [1.1.4] - 2023-10-13
 
 ### Fixed